+++ src/MenuGame.cpp
@@ -176,7 +176,7 @@ namespace Ui {
                                     fg = e->getFgColorCode();
                                     bg = e->getBgColorCode();
                                 }
-                                symbol = e->getIcon(p, (int)tick, currentWorld->currentLevel);
+                                symbol = e->getIcon(p, tick, currentWorld->currentLevel);
                             }
 
                         }
@@ -216,7 +216,7 @@ namespace Ui {
             targetPosition += p;
         }if(mode == modeInterectChoose){
 
-            Item* i = *useItem!=-1?(currentWorld->currentPlayer->inventory[(size_t)*useItem]):currentWorld->currentPlayer->getActiveWeapon();
+            Item* i = *useItem!=-1?(currentWorld->currentPlayer->inventory[*useItem]):currentWorld->currentPlayer->getActiveWeapon();
             
             timePassed += currentWorld->currentPlayer->interact(currentWorld->currentLevel, currentWorld->currentPlayer->pos+p, false, i);
 
@@ -421,7 +421,7 @@ namespace Ui {
 
             mvprintw(a, gameArea.x+1, "HP: %d/%d", hp, maxHp);
             Ui::setColor((hp<(maxHp/3*2))?((hp<(maxHp/3))?C_LIGHT_RED:C_LIGHT_YELLOW):C_LIGHT_GREEN);
-            printw(" %s", StringUtility::makeBar(hp, maxHp, (size_t)(terminalSize.x - getcurx(stdscr) - 2)).c_str());
+            printw(" %s", StringUtility::makeBar(hp, maxHp, (terminalSize.x - getcurx(stdscr) - 2)).c_str());
             Ui::setColor(C_WHITE);
 
             a++;
@@ -431,7 +431,7 @@ namespace Ui {
 
             mvprintw(a, gameArea.x+1, "MP: %d/%d", mp, maxMp);
             Ui::setColor(C_LIGHT_BLUE);
-            printw(" %s", StringUtility::makeBar(mp, maxMp, (size_t)(terminalSize.x - getcurx(stdscr) - 2)).c_str());
+            printw(" %s", StringUtility::makeBar(mp, maxMp, (terminalSize.x - getcurx(stdscr) - 2)).c_str());
             Ui::setColor(C_WHITE);
 
             a++;
@@ -492,7 +492,7 @@ namespace Ui {
                         const int maxHp = Math::roundToInt(aiEntity->getMaxHp());
                         mvprintw(a, gameArea.x+1, " HP: %d/%d", hp, maxHp);
                         Ui::setColor((hp<(maxHp/3*2))?((hp<(maxHp/3))?C_LIGHT_RED:C_LIGHT_YELLOW):C_LIGHT_GREEN);
-                        printw(" %s", StringUtility::makeBar(hp, maxHp, (size_t)(terminalSize.x - getcurx(stdscr) - 2)).c_str());
+                        printw(" %s", StringUtility::makeBar(hp, maxHp, (terminalSize.x - getcurx(stdscr) - 2)).c_str());
                         Ui::setColor(C_WHITE);
                         a++;
                     }
